Refactor: Move blazeCampaigns under blaze package
