+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityNavigator.kt
@@ -2,11 +2,11 @@ package org.wordpress.android.ui
 
 import android.content.Context
 import android.content.Intent
-import org.wordpress.android.ui.blazeCampaigns.ARG_EXTRA_BLAZE_CAMPAIGN_PAGE
-import org.wordpress.android.ui.blazeCampaigns.BlazeCampaignPage
-import org.wordpress.android.ui.blazeCampaigns.BlazeCampaignParentActivity
-import org.wordpress.android.ui.blazeCampaigns.campaigndetail.CampaignDetailPageSource
-import org.wordpress.android.ui.blazeCampaigns.campaignlisting.CampaignListingPageSource
+import org.wordpress.android.ui.blaze.blazeCampaigns.ARG_EXTRA_BLAZE_CAMPAIGN_PAGE
+import org.wordpress.android.ui.blaze.blazeCampaigns.BlazeCampaignPage
+import org.wordpress.android.ui.blaze.blazeCampaigns.BlazeCampaignParentActivity
+import org.wordpress.android.ui.blaze.blazeCampaigns.campaigndetail.CampaignDetailPageSource
+import org.wordpress.android.ui.blaze.blazeCampaigns.campaignlisting.CampaignListingPageSource
 import javax.inject.Inject
 import javax.inject.Singleton
