+++ WordPress/src/main/java/org/wordpress/android/ui/blaze/blazeCampaigns/campaigndetail/CampaignDetailViewModel.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.blazeCampaigns.campaigndetail
+package org.wordpress.android.ui.blaze.blazeCampaigns.campaigndetail
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
